@@ -1,26 +1,50 @@
 {
-  "name": "fashion-clothing",
+  "name": "crwn-clothing",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "firebase": "^7.14.6",
-    "node-sass": "^7.0.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1"
+    "@stripe/react-stripe-js": "1.7.0",
+    "@stripe/stripe-js": "1.25.0",
+    "@testing-library/jest-dom": "5.16.1",
+    "@testing-library/react": "12.1.2",
+    "@testing-library/user-event": "13.5.0",
+    "@types/jest": "27.4.1",
+    "@types/node": "17.0.23",
+    "@types/react": "17.0.43",
+    "@types/react-dom": "17.0.14",
+    "@types/redux-logger": "3.0.9",
+    "@types/styled-components": "5.1.24",
+    "dotenv": "16.0.0",
+    "firebase": "9.6.6",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "7.2.6",
+    "react-router-dom": "6",
+    "react-scripts": "^5.0.1",
+    "redux": "4.1.2",
+    "redux-logger": "3.0.6",
+    "redux-persist": "6.0.0",
+    "redux-saga": "1.1.3",
+    "redux-thunk": "2.4.1",
+    "reselect": "4.1.5",
+    "sass": "1.49.7",
+    "stripe": "8.210.0",
+    "styled-components": "5.3.3",
+    "typed-redux-saga": "1.4.0",
+    "typescript": "4.6.3",
+    "web-vitals": "2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-scripts --openssl-legacy-provider start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -33,5 +57,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "babel-plugin-macros": "3.1.0"
   }
 }
